Extract DB creation into separate recipe
Because it's possible to have a separate machine with MySQL only,
it doesn't make sense installing all the tomcats & nginx there, thus
DB-related logic was extracted into separate recipe.

Note, that in Packer it's important what order we defined in the run_list
since the latest cookbook will override the attributes of the preceding
cookbooks.